Update viewData and viewParams support to handle passing data or params for specific views. To specify that data or a param is for a specific view use {viewParams:{views:{viewDest:{paramName:paramValue}}}}, the same format works for viewData.  To set params or data available for all views being transitioned to, put the params outside of the views like this: {viewParams:{paramName:paramValue}}
Update to PR to move foundView function in hash.js into the closure since it will not be called outside of hash.  This PR is to update viewData and viewParams support to handle passing data or params for specific views. To specify that data or a param is for a specific view use {viewParams:{views:{viewDest:{paramName:paramValue}}}}, the same format works for viewData.  To set params or data available for all views being transitioned to, put the params outside of the views like this: {viewParams:{paramName:paramValue}}