Updates to dapp for the History controller support including updates to pass viewParams and viewData with params and data for specific views, support to hide a view on browser back if it was just shown on the current view, and tests for all of the history support, as well as new config options for alwaysUseDefaultView and hideUrlHash
I noticed that I left onclick in one of the history test templates, update it to on-click

Updates to remove dojo/_base references from the History controller, and to move History controller out of the delite specific controllers folder into the controllers folder.

Change event name from dapp-finishedTransition to dapp-finished-transition.